backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
faiss-cpu
sentence-transformers
numpy
tiktoken
